Move fixtures to separate class
